Move 'Configuration' class to 'wyopcl' package
