+++ modules/wyopcl/src/wyopcl/Configuration.java
@@ -1,4 +1,4 @@
-package wyopcl.translator;
+package wyopcl;
 
 import java.util.Properties;
