+++ modules/wyopcl/src/wyopcl/translator/bound/CFGraph.java
@@ -11,8 +11,8 @@ import java.util.List;
 import wyil.lang.Code;
 import wyil.lang.Codes;
 import wyil.lang.Type;
+import wyopcl.Configuration;
 import wyopcl.translator.AnalyzerHelper;
-import wyopcl.translator.Configuration;
 import wyopcl.translator.TranslatorHelper;
 import wyopcl.translator.bound.BasicBlock.BlockType;
 import wyopcl.translator.bound.constraint.Constraint;
