+++ modules/wyopcl/src/wyopcl/translator/generator/AbstractCodeGenerator.java
@@ -51,7 +51,7 @@ import wyil.lang.Type;
 import wyil.lang.WyilFile;
 import wyil.lang.WyilFile.Constant;
 import wyil.lang.WyilFile.FunctionOrMethod;
-import wyopcl.translator.Configuration;
+import wyopcl.Configuration;
 
 /**
  * Defines the functions for code generator.
