+++ modules/wyopcl/src/wyopcl/translator/generator/CodeStores.java
@@ -12,7 +12,7 @@ import wyil.lang.Code;
 import wyil.lang.Codes;
 import wyil.lang.Type;
 import wyil.lang.WyilFile.FunctionOrMethod;
-import wyopcl.translator.Configuration;
+import wyopcl.Configuration;
 
 /**
  * Stores the generated code and provides common functionality for code generator.
