+++ modules/wyopcl/src/wyopcl/translator/symbolic/pattern/Pattern.java
@@ -10,7 +10,7 @@ import wyil.lang.Code;
 import wyil.lang.Codes;
 import wyil.lang.Type;
 import wyil.lang.WyilFile.FunctionOrMethod;
-import wyopcl.translator.Configuration;
+import wyopcl.Configuration;
 import wyopcl.translator.symbolic.pattern.expression.ExprFactory;
 /**
  * The abstract base class to define all the commonly used fields and methods for all its subclass, including the assert_label, data-driven storage and
