+++ modules/wyopcl/src/wyopcl/translator/symbolic/pattern/WhileLoopDecrPattern.java
@@ -4,7 +4,7 @@ import java.util.List;
 import wyil.lang.Code;
 import wyil.lang.Type;
 import wyil.lang.WyilFile.FunctionOrMethod;
-import wyopcl.translator.Configuration;
+import wyopcl.Configuration;
 /**
  * The final class that implemented the while-loop pattens, as follows:
  * <ul>
