+++ modules/wyopcl/src/wyopcl/translator/symbolic/pattern/WhileLoopPattern.java
@@ -10,7 +10,7 @@ import wyil.lang.Codes.Loop;
 import wyil.lang.Type;
 import wyil.lang.Codes.Comparator;
 import wyil.lang.WyilFile.FunctionOrMethod;
-import wyopcl.translator.Configuration;
+import wyopcl.Configuration;
 import wyopcl.translator.symbolic.pattern.expression.LinearExpr;
 
 /**
